Further optimise day 23
Switch to recursive DFS to avoid having to clone the state. Also
avoid hash key joining and splitting in part 2 which makes quite
a large performance difference.

In part 1 setting the max path to a negative number works but in
part 2 it causes the solution to be marginally out. Not 100% sure
why but it must be related to the weights not all being 1.